packages/backend/requirements.txt
python-ulid==3.0.0
Flask==3.0.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.1.0
SQLAlchemy==2.0.40
typing-extensions==4.13.0
python-dotenv==1.0.1
pytest==8.0.2
black==24.2.0
isort==5.13.2
mypy==1.8.0
gunicorn==21.2.0
bcrypt>=4.3.0
PyJWT>=2.6.0
pymysql>=1.0.2
cryptography>=36.0.0
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.4.2
pydantic-core==2.10.1
python-jose>=3.4.0
passlib>=1.7.4
types-python-jose>=3.3.0
types-passlib>=1.7.7.20240104
pydantic_settings==2.9.1